LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/open-send-during-abort-expected.txt


FAIL XMLHttpRequest: open() during abort() assert_array_equals: property 2, expected 1 but got 0
